requirements.txt
# Python 3.10+ compatible requirements
# Install: pip install -r requirements.txt
#
# OpenSim 4.4 installs via pip -- no conda channel needed:
#   pip install opensim
# or via conda:
#   conda install -c opensim-org opensim

# Core simulation
osim-rl>=4.0

# RL framework (replaces keras-rl)
stable-baselines3>=2.3.0
torch>=2.2.0

# Gym interface
gymnasium>=0.29.0

# Numerics / plotting
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
scipy>=1.11.0

# Utilities
tqdm>=4.66.0
pickle5; python_version < '3.8'
